requirements.txt
python-dotenv
langchain-google-genai
langchain-core
google-genai
requests
yfinance
pandas
rich
